// src/security/dto/facebook-credentials.dto.ts
export class FacebookCredentials
{
    accessToken: string;

    accessTokenSource: string;

    applicationID: string;

    dataAccessExpirationTime: number;

    declinedPermissions: Array<string>;

    expirationTime: number;

    expiredPermissions: Array<string>;

    lastRefreshTime: number;

    permissions: Array<string>;

    userID: string;
}